improvement: refoldering swagger
