+++ libs/controllers/src/student/index.ts
@@ -13,15 +13,13 @@ import {
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
 import {
-  CreateStudentSwagger,
   CreateStudentZodSchema,
   JwtAuthGuard,
   TReqToken,
-  UpdateStudentSwagger,
   UpdateStudentZodSchema,
   ZodValidationPipe,
 } from '@uninus/entities';
-import { StudentService } from '@uninus/services';
+import { StudentService, UpdateStudentSwagger } from '@uninus/services';
 import {
   ApiResponse,
   ApiTags,
@@ -29,6 +27,7 @@ import {
   ApiOperation,
   ApiBearerAuth,
 } from '@nestjs/swagger';
+import { CreateStudentSwagger } from '@uninus/services';
 
 @Controller('student')
 @ApiTags('Student')
