+++ libs/entities/src/schemas/auth-schemas/login/index.ts
@@ -1,2 +1,1 @@
 export * from './index.schema';
-export * from './index.swagger';
