+++ libs/entities/src/schemas/auth-schemas/logout/index.ts
@@ -1,2 +1,1 @@
 export * from './index.schema';
-export * from './index.swagger';
