+++ libs/entities/src/schemas/student-schemas/create-student/index.ts
@@ -1,2 +1,1 @@
 export * from './index.schema';
-export * from './index.swagger';
