+++ libs/entities/src/schemas/user-schemas/create-user/index.ts
@@ -1,2 +1,1 @@
 export * from './index.schema';
-export * from './index.swagger';
