+++ libs/services/src/index.ts
@@ -6,3 +6,4 @@ export * from './student';
 export * from './user';
 export * from './cloudinary';
 export * from './location';
+export * from './swagger';
