+++ libs/services/src/swagger/auth-swagger/forgot-password/index.swagger.ts
