+++ libs/services/src/swagger/auth-swagger/index.ts
@@ -1,0 +1,8 @@
+export * from './forgot-password';
+export * from './login';
+export * from './logout';
+export * from './register';
+export * from './req-user';
+export * from './resend-otp';
+export * from './reset-password';
+export * from './verify';
