+++ libs/services/src/swagger/auth-swagger/login/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
