+++ libs/services/src/swagger/auth-swagger/logout/index.swagger.ts
