+++ libs/services/src/swagger/auth-swagger/logout/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
