+++ libs/services/src/swagger/auth-swagger/register/index.swagger.ts
