+++ libs/services/src/swagger/auth-swagger/register/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
