+++ libs/services/src/swagger/auth-swagger/resend-otp/index.swagger.ts
