+++ libs/services/src/swagger/auth-swagger/verify/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
