+++ libs/services/src/swagger/index.ts
@@ -1,0 +1,3 @@
+export * from './auth-swagger';
+export * from './student-swagger';
+export * from './user-swagger';
