+++ libs/services/src/swagger/student-swagger/index.ts
@@ -1,0 +1,2 @@
+export * from './create-student';
+export * from './update-student';
