+++ libs/services/src/swagger/student-swagger/update-student/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
