+++ libs/services/src/swagger/user-swagger/create-user/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
