+++ libs/services/src/swagger/user-swagger/index.ts
@@ -1,0 +1,2 @@
+export * from './create-user';
+export * from './update-user';
