+++ libs/services/src/swagger/user-swagger/update-user/index.ts
@@ -1,0 +1,1 @@
+export * from './index.swagger';
